ChessAI/requirements.txt
aiofiles==22.1.0
aiosqlite==0.19.0
alabaster==0.7.13
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.3
asttokens==2.2.1
atomicwrites==1.4.1
attrs==23.1.0
autopep8==2.0.2
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==23.7.0
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.6
cloudpickle==2.2.1
comm==0.1.3
contourpy==1.0.7
cookiecutter==2.3.0
cryptography==41.0.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.6
docstring-to-markdown==0.12
docutils==0.20.1
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
flake8==6.0.0
fonttools==4.39.3
fqdn==1.5.1
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
inflection==0.5.1
iniconfig==2.0.0
intervaltree==3.1.0
ipdb==0.13.13
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
isoduration==20.11.0
isort==5.12.0
jaraco.classes==3.3.0
jedi==0.18.2
jeepney==0.8.0
jellyfish==1.0.0
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
keyring==24.2.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.5
more-itertools==10.1.0
mypy-extensions==1.0.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
numpydoc==1.5.0
packaging==23.1
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.15.1
pylint==2.17.2
pylint-venv==3.0.2
pyls-spyder==0.4.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-lsp-black==1.3.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.4
python-slugify==8.0.1
pytoolconfig==1.2.5
pytz==2023.3
pyxdg==0.28
PyYAML==6.0
pyzmq==25.0.2
QDarkStyle==3.1
qstylizer==0.2.2
QtAwesome==1.2.3
qtconsole==5.4.3
QtPy==2.3.1
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rope==1.9.0
Rtree==1.0.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4.1
Sphinx==7.1.2
sphinxcontrib-applehelp==1.0.6
sphinxcontrib-devhelp==1.0.4
sphinxcontrib-htmlhelp==2.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.5
sphinxcontrib-serializinghtml==1.1.7
spyder==5.4.4
spyder-kernels==2.4.4
stack-data==0.6.2
terminado==0.17.1
text-unidecode==1.3
textdistance==4.5.0
threadpoolctl==3.1.0
three-merge==0.1.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.7
tornado==6.3
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.8.0
uri-template==1.2.0
urllib3==1.26.15
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
whatthepatch==1.0.5
wrapt==1.15.0
wurlitzer==3.0.3
y-py==0.5.9
yapf==0.40.1
ypy-websocket==0.8.2
zipp==3.16.2
